feat: add method to get guild badges
